First checkout

+++ basenode.cpp
@@ -1,29 +1,23 @@
 #include "basenode.h"
 
-
 void BaseNode::updateCircleTable(float radius, float x, float y)
 {
     QSGGeometry *g = geometry();
     g->allocate(unitCircleTableX.size());
     QSGGeometry::TexturedPoint2D *v = g->vertexDataAsTexturedPoint2D();
 
-
-    for(int i=0; i<unitCircleTableX.size(); i++){
+    for(size_t i=0; i<unitCircleTableX.size(); i++){
         v[i].set(unitCircleTableX[i]*radius+x, unitCircleTableY[i]*radius+y,
                  (unitCircleTableX[i]+1)/2,(unitCircleTableY[i]+1)/2);
-
     }
 }
 
-void BaseNode::initCircleTable()
-{
+void BaseNode::initCircleTable(){
 
     const unsigned HALF_PREC = 10;
     const float angInc = M_PI / static_cast<float>(HALF_PREC);
     const float cosInc = cos(angInc);
     const float sinInc = sin(angInc);
-    //unsigned coordIdx = 0;
-
 
     float xc = 1.0f;
     float yc = 0.0f;
@@ -34,7 +28,7 @@ void BaseNode::initCircleTable()
     unitCircleTableX.push_back(xc);
     unitCircleTableY.push_back(yc);
 
-    for (int iAng= 1; iAng<HALF_PREC ; ++iAng)  {
+    for (unsigned int iAng=1; iAng<HALF_PREC; ++iAng)  {
         float xcNew = cosInc * xc - sinInc * yc;
         yc = sinInc * xc + cosInc * yc;
         xc = xcNew;
